Refactored generators
