+++ src/lib/cli/cmds/generate_cmds/resolver/create-args.ts
@@ -1,0 +1,16 @@
+import titleize from '../../../helpers/titleize'
+
+export const createArgs = (name: string) => {
+  return (`import {
+  ArgsType,
+  Field,
+  ID
+} from 'type-graphql'
+
+@ArgsType()
+export class Get${titleize(name)}Args {
+  @Field(type => ID)
+  id: string
+}
+`)
+}
